Çizerken yanlış son nokta çizimi kaybettirmeden geri alınıyor (C-02)
ÇİZGİ, ÇOKLUÇİZGİ, ALAN ve SPLINE nokta beklerken ⌫, Ctrl+Z, Geri Al ve
komut satırına yazılan G/GERİ/U yalnız son noktayı geri alıyor; çalışmanın
geri kalanı tuvalde kalıyor ve komut o noktayı yeniden istiyor. Eskiden ⌫
çalışmayı yazıp SİL'i başlatıyor, Ctrl+Z ve U çalışmayı yazıp tamamını geri
alıyordu.

- Session::retract / Prompt::can_retract / Context::took_back; kelimeler
  command::asks_retract'ta tek yerde (GERİAL'in kendi adları dahil).
- ÇİZGİ çalışmayı bitince yazıyor (ÇOKLUÇİZGİ gibi): geri alınan parça
  anahtar harcamıyor, günlük aynı anahtarlarla yeniden oynuyor.
- Yakalama çalışmanın bekleyen köşelerini (UÇ) ve kenarlarını (ORTA, YAKIN,
  DİK, KESİŞİM) buluyor; komutla birlikte yazılan noktalar bunlara çekilmiyor
  (gerçek pencere probunun yakaladığı çökme).
- Tuvaldeki yakalama önizlemesi izleme işaretlerini ve çalışmayı komutla
  aynı çağrıyla hesaba katıyor.
- Boş komut satırında Ctrl+Z artık satırın metin geçmişine değil GERİAL'e
  gidiyor; boş satırda ⌫ son noktayı geri alıyor.

Kanıt: tests/unit/test_draw_methods.cpp, PROOF: geri alınan köşe hiçbir
yolda kalmaz, gerçek pencerede KENTOS_REALMOUSE_PROBE 7b.

+++ tests/unit/test_command.cpp
@@ -6149,7 +6149,7 @@ core::Point2 ring_vertex(const core::Document& doc, std::uint64_t key, std::size
     const core::EntityId e = doc.slot_of(static_cast<core::EntityKey>(key));
     REQUIRE(e != core::kNoEntity);
     const core::RingSpan span = doc.geometry().rings_of(doc.entities().slot[e]);
-    return doc.geometry().vertex(span.first, at);
+    return doc.geometry().vertex(span.first, static_cast<std::uint32_t>(at));
 }
 
 core::KindId kind_of(const core::Document& doc, std::uint64_t key)
